tests: fatal: make sure the illegal insn occupies a full word
... and is properly aligned. This may make a difference on
64-bit targets.

+++ tests/kernel/fatal/src/main.c
@@ -63,7 +63,7 @@ void alt_thread1(void)
 	 * and xtensa
 	 */
 	{
-		int illegal = 0;
+		long illegal = 0;
 		((void(*)(void))&illegal)();
 	}
 #endif
